Added support for oblique gesture swipes.
To add the support the config class was modified to include the new
gestures types: left_up, right_up, left_down, right_down.
Also changed the way the commands are saved in the config class, so
there won't be a need for a lot of if statements to handle 4 finger
swipes and 3 finger swipes.
Finally, the handle_swipe_event_without_coords was modified to detect
the oblique swipes.